requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
azure-core==1.30.2
azure-storage-blob==12.20.0
boto3==1.34.142
botocore==1.34.142
build==1.2.1
CacheControl==0.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
crispy-bootstrap5==2024.2
cryptography==42.0.8
dataclasses-json==0.6.7
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0.6
django-crispy-forms==2.1
dulwich==0.21.7
fastapi==0.110.0
fastapi-jwt-auth==0.5.0
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.50.0
fpdf2==2.7.8
frozenlist==1.4.1
fsspec==2024.6.1
grpcio==1.63.0
grpcio-tools==1.63.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==24.3.1
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.12
langchain-text-splitters==0.2.2
langsmith==0.1.85
mangum==0.17.0
MarkupSafe==2.1.5
marshmallow==3.21.3
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
openai==1.14.2
orjson==3.10.6
packaging==24.1
pandas==2.2.1
param==2.1.1
pexpect==4.9.0
Pillow==9.5.0
pkginfo==1.11.1
platformdirs==4.2.2
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
prisma==0.13.1
protobuf==5.27.2
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==1.7.1
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
stability-sdk==0.8.6
starlette==0.36.3
sympy==1.13.0
tenacity==8.5.0
tomlkit==0.12.5
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.2
trove-classifiers==2024.7.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.29.0
virtualenv==20.26.3
Werkzeug==3.0.1
xattr==1.1.0
yarl==1.9.4
